Move elements and attributes into html package
This makes it easier to use dot-imports.

Also updated the readme and examples with new usage.
