--- attr/attributes_test.go
@@ -1,32 +1,0 @@
-package attr_test
-
-import (
-	"testing"
-
-	g "github.com/maragudk/gomponents"
-	"github.com/maragudk/gomponents/assert"
-	"github.com/maragudk/gomponents/attr"
-)
-
-func TestClasses(t *testing.T) {
-	t.Run("given a map, returns sorted keys from the map with value true", func(t *testing.T) {
-		assert.Equal(t, ` class="boheme-hat hat partyhat"`, attr.Classes{
-			"boheme-hat": true,
-			"hat":        true,
-			"partyhat":   true,
-			"turtlehat":  false,
-		})
-	})
-
-	t.Run("renders as attribute in an element", func(t *testing.T) {
-		e := g.El("div", attr.Classes{"hat": true})
-		assert.Equal(t, `<div class="hat"></div>`, e)
-	})
-
-	t.Run("also works with fmt", func(t *testing.T) {
-		a := attr.Classes{"hat": true}
-		if a.String() != ` class="hat"` {
-			t.FailNow()
-		}
-	})
-}
