--- attr/boolean.go
@@ -1,45 +1,0 @@
-package attr
-
-import (
-	g "github.com/maragudk/gomponents"
-)
-
-func Async() g.Node {
-	return g.Attr("async")
-}
-
-func AutoFocus() g.Node {
-	return g.Attr("autofocus")
-}
-
-func AutoPlay() g.Node {
-	return g.Attr("autoplay")
-}
-
-func Controls() g.Node {
-	return g.Attr("controls")
-}
-
-func Defer() g.Node {
-	return g.Attr("defer")
-}
-
-func Disabled() g.Node {
-	return g.Attr("disabled")
-}
-
-func Multiple() g.Node {
-	return g.Attr("multiple")
-}
-
-func ReadOnly() g.Node {
-	return g.Attr("readonly")
-}
-
-func Required() g.Node {
-	return g.Attr("required")
-}
-
-func Selected() g.Node {
-	return g.Attr("selected")
-}
