+++ components/documents.go
@@ -3,8 +3,7 @@ package components
 
 import (
 	g "github.com/maragudk/gomponents"
-	"github.com/maragudk/gomponents/attr"
-	"github.com/maragudk/gomponents/el"
+	. "github.com/maragudk/gomponents/html"
 )
 
 // DocumentProps for HTML5.
@@ -21,21 +20,21 @@ type DocumentProps struct {
 func HTML5(p DocumentProps) g.NodeFunc {
 	var lang, description g.Node
 	if p.Language != "" {
-		lang = attr.Lang(p.Language)
+		lang = Lang(p.Language)
 	}
 	if p.Description != "" {
-		description = el.Meta(attr.Name("description"), attr.Content(p.Description))
+		description = Meta(Name("description"), Content(p.Description))
 	}
-	return el.Document(
-		el.HTML(lang,
-			el.Head(
-				el.Meta(attr.Charset("utf-8")),
-				el.Meta(attr.Name("viewport"), attr.Content("width=device-width, initial-scale=1")),
-				el.Title(p.Title),
+	return Document(
+		HTML(lang,
+			Head(
+				Meta(Charset("utf-8")),
+				Meta(Name("viewport"), Content("width=device-width, initial-scale=1")),
+				TitleEl(p.Title),
 				description,
 				g.Group(p.Head),
 			),
-			el.Body(g.Group(p.Body)),
+			Body(g.Group(p.Body)),
 		),
 	)
 }
