--- el/simple.go
@@ -1,257 +1,0 @@
-package el
-
-import (
-	g "github.com/maragudk/gomponents"
-)
-
-func Address(children ...g.Node) g.NodeFunc {
-	return g.El("address", children...)
-}
-
-func Area(children ...g.Node) g.NodeFunc {
-	return g.El("area", children...)
-}
-
-func Article(children ...g.Node) g.NodeFunc {
-	return g.El("article", children...)
-}
-
-func Aside(children ...g.Node) g.NodeFunc {
-	return g.El("aside", children...)
-}
-
-func Audio(children ...g.Node) g.NodeFunc {
-	return g.El("audio", children...)
-}
-
-func Base(children ...g.Node) g.NodeFunc {
-	return g.El("base", children...)
-}
-
-func BlockQuote(children ...g.Node) g.NodeFunc {
-	return g.El("blockquote", children...)
-}
-
-func Body(children ...g.Node) g.NodeFunc {
-	return g.El("body", children...)
-}
-
-func Br(children ...g.Node) g.NodeFunc {
-	return g.El("br", children...)
-}
-
-func Button(children ...g.Node) g.NodeFunc {
-	return g.El("button", children...)
-}
-
-func Canvas(children ...g.Node) g.NodeFunc {
-	return g.El("canvas", children...)
-}
-
-func Cite(children ...g.Node) g.NodeFunc {
-	return g.El("cite", children...)
-}
-
-func Code(children ...g.Node) g.NodeFunc {
-	return g.El("code", children...)
-}
-
-func Col(children ...g.Node) g.NodeFunc {
-	return g.El("col", children...)
-}
-
-func ColGroup(children ...g.Node) g.NodeFunc {
-	return g.El("colgroup", children...)
-}
-
-func Data(children ...g.Node) g.NodeFunc {
-	return g.El("data", children...)
-}
-
-func DataList(children ...g.Node) g.NodeFunc {
-	return g.El("datalist", children...)
-}
-
-func Details(children ...g.Node) g.NodeFunc {
-	return g.El("details", children...)
-}
-
-func Dialog(children ...g.Node) g.NodeFunc {
-	return g.El("dialog", children...)
-}
-
-func Div(children ...g.Node) g.NodeFunc {
-	return g.El("div", children...)
-}
-
-func Dl(children ...g.Node) g.NodeFunc {
-	return g.El("dl", children...)
-}
-
-func Embed(children ...g.Node) g.NodeFunc {
-	return g.El("embed", children...)
-}
-
-func FieldSet(children ...g.Node) g.NodeFunc {
-	return g.El("fieldset", children...)
-}
-
-func Figure(children ...g.Node) g.NodeFunc {
-	return g.El("figure", children...)
-}
-
-func Footer(children ...g.Node) g.NodeFunc {
-	return g.El("footer", children...)
-}
-
-func Head(children ...g.Node) g.NodeFunc {
-	return g.El("head", children...)
-}
-
-func Header(children ...g.Node) g.NodeFunc {
-	return g.El("header", children...)
-}
-
-func HGroup(children ...g.Node) g.NodeFunc {
-	return g.El("hgroup", children...)
-}
-
-func Hr(children ...g.Node) g.NodeFunc {
-	return g.El("hr", children...)
-}
-
-func HTML(children ...g.Node) g.NodeFunc {
-	return g.El("html", children...)
-}
-
-func IFrame(children ...g.Node) g.NodeFunc {
-	return g.El("iframe", children...)
-}
-
-func Legend(children ...g.Node) g.NodeFunc {
-	return g.El("legend", children...)
-}
-
-func Li(children ...g.Node) g.NodeFunc {
-	return g.El("li", children...)
-}
-
-func Link(children ...g.Node) g.NodeFunc {
-	return g.El("link", children...)
-}
-
-func Main(children ...g.Node) g.NodeFunc {
-	return g.El("main", children...)
-}
-
-func Menu(children ...g.Node) g.NodeFunc {
-	return g.El("menu", children...)
-}
-
-func Meta(children ...g.Node) g.NodeFunc {
-	return g.El("meta", children...)
-}
-
-func Meter(children ...g.Node) g.NodeFunc {
-	return g.El("meter", children...)
-}
-
-func Nav(children ...g.Node) g.NodeFunc {
-	return g.El("nav", children...)
-}
-
-func NoScript(children ...g.Node) g.NodeFunc {
-	return g.El("noscript", children...)
-}
-
-func Object(children ...g.Node) g.NodeFunc {
-	return g.El("object", children...)
-}
-
-func Ol(children ...g.Node) g.NodeFunc {
-	return g.El("ol", children...)
-}
-
-func OptGroup(children ...g.Node) g.NodeFunc {
-	return g.El("optgroup", children...)
-}
-
-func P(children ...g.Node) g.NodeFunc {
-	return g.El("p", children...)
-}
-
-func Param(children ...g.Node) g.NodeFunc {
-	return g.El("param", children...)
-}
-
-func Picture(children ...g.Node) g.NodeFunc {
-	return g.El("picture", children...)
-}
-
-func Pre(children ...g.Node) g.NodeFunc {
-	return g.El("pre", children...)
-}
-
-func Script(children ...g.Node) g.NodeFunc {
-	return g.El("script", children...)
-}
-
-func Section(children ...g.Node) g.NodeFunc {
-	return g.El("section", children...)
-}
-
-func Source(children ...g.Node) g.NodeFunc {
-	return g.El("source", children...)
-}
-
-func Span(children ...g.Node) g.NodeFunc {
-	return g.El("span", children...)
-}
-
-func Style(children ...g.Node) g.NodeFunc {
-	return g.El("style", children...)
-}
-
-func Summary(children ...g.Node) g.NodeFunc {
-	return g.El("summary", children...)
-}
-
-func SVG(children ...g.Node) g.NodeFunc {
-	return g.El("svg", children...)
-}
-
-func Table(children ...g.Node) g.NodeFunc {
-	return g.El("table", children...)
-}
-
-func TBody(children ...g.Node) g.NodeFunc {
-	return g.El("tbody", children...)
-}
-
-func Td(children ...g.Node) g.NodeFunc {
-	return g.El("td", children...)
-}
-
-func TFoot(children ...g.Node) g.NodeFunc {
-	return g.El("tfoot", children...)
-}
-
-func Th(children ...g.Node) g.NodeFunc {
-	return g.El("th", children...)
-}
-
-func THead(children ...g.Node) g.NodeFunc {
-	return g.El("thead", children...)
-}
-
-func Tr(children ...g.Node) g.NodeFunc {
-	return g.El("tr", children...)
-}
-
-func Ul(children ...g.Node) g.NodeFunc {
-	return g.El("ul", children...)
-}
-
-func Wbr(children ...g.Node) g.NodeFunc {
-	return g.El("wbr", children...)
-}
